Move pair loop rng tests into separate file.
